requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
scikit-learn
scipy
dahuffman
tqdm
Pillow
joblib
opencv-python
scikit-image